Update for Julia 1.0 (#51)
* Updates for Julia 1.0

Due to new mapreduce syntax and new iteration protocol.
Renamed variable state[12] to elstate[12] because iterate returns
both element and state (or nothing).

* Drop support for Julia 0.6

As the new iteration protocol is not supported in Julia 0.6 nor in
Compat, set minimum required Julia version to 0.7 and remove the
dependency on Compat. Update README to reflect this change. Update CI
scripts accordingly. Rewrote the appveyor script based on the JuliaLang
Example repo.