Change `conj` flags from `Symbol` to `Bool` (#176)
This PR simplifies the usage of conj flags by replacing the symbols with booleans. This is especially useful in AD, where comparing and negating with booleans is more convenient.

Importantly, as a Bool is a subtype of Number, this introduces some method ambiguities for the functions syntax, which is resolved by restricting the types of the labels to be a Tuple or AbstractVector of Int, Symbol or Char.